+++ db/migrate/20200127112223_delete_empty_guilds.rb
@@ -1,0 +1,5 @@
+class DeleteEmptyGuilds < ActiveRecord::Migration[5.2]
+  def change
+    Guild.where(characters_count: 0).destroy_all
+  end
+end